safari-600.7-branch/LayoutTests/js/dom/native-error-prototype-expected.txt

This is a test case for bugs 55346, 70889, and 75452.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ({}).toString.call(Error.prototype) is "[object Error]"
PASS ({}).toString.call(RangeError.prototype) is "[object Error]"
PASS err.toString() is "message"
PASS err.hasOwnProperty('message') is false
PASS err.hasOwnProperty('message') is false
PASS successfullyParsed is true

TEST COMPLETE
